network-vuln-scanner/requirements.txt
python-nmap
schedule
jinja2
smtplib-ssl
psutil
colorama